Bridging FE and STE

+++ cmd/copy.go
@@ -90,7 +90,7 @@ func init() {
 	cpCmd.PersistentFlags().BoolVar(&commandLineInput.WithSnapshots, "with-snapshots", false, "Filter: Include the snapshots. Only valid when the source is blobs.")
 
 	// options
-	cpCmd.PersistentFlags().Uint16Var(&commandLineInput.BlockSize, "block-size", 0, "Use this block size when uploading to Azure Storage.")
+	cpCmd.PersistentFlags().Uint32Var(&commandLineInput.BlockSize, "block-size", 0, "Use this block size when uploading to Azure Storage.")
 	cpCmd.PersistentFlags().StringVar(&commandLineInput.BlobType, "blob-type", "block", "Upload to Azure Storage using this blob type.")
 	cpCmd.PersistentFlags().StringVar(&commandLineInput.BlobTier, "blob-tier", "", "Upload to Azure Storage using this blob tier.")
 	cpCmd.PersistentFlags().StringVar(&commandLineInput.Metadata, "metadata", "", "Upload to Azure Storage with these key-value pairs as metadata.")
